Update README.md
short update for a project readme whose source actually currently resides in an GCP repo. We've an IP lawyer search in progress to help us properly sort out original works from every bit of shareable tech we have in addition to all of the open source code that helps make plenty of our things and stuff.
This commit is long.
I'm so sleepy! Okay, update the repo in a month. Will try to provide code relevant to this effort.